requirements.txt
# Outras dependencias essenciais
numpy<2.1.0
g4f[all]
yt-dlp
ffmpeg-python
whisperx
mediapipe
google-genai
insightface
gradio
opencv-python
psutil
fastapi
uvicorn
deep-translator
tqdm
edge-tts
python-dotenv
openai
anthropic
google-generativeai
